app/requirements.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.4.2
google-api-core==2.5.0
google-api-python-client==2.36.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-language==2.3.2
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
h11==0.13.0
httplib2==0.20.4
idna==3.3
jieba3k==0.35.1
joblib==1.1.0
kafka-python==2.0.2
lxml==4.7.1
newspaper3k==0.2.8
nltk==3.6.7
outcome==1.1.0
Pillow==9.0.1
proto-plus==1.19.9
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
requests-file==1.5.1
rsa==4.8
selenium==4.1.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
tinysegmenter==0.3
tldextract==3.1.2
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.8
webdriver-manager==3.5.2
wsproto==1.0.0
